migrate memory governor to a new reusable header
