+++ src/include/duckdb/common/serializer/async_write_queue.hpp
@@ -13,6 +13,7 @@
 #include "duckdb/common/error_data.hpp"
 #include "duckdb/common/mutex.hpp"
 #include "duckdb/common/optional_ptr.hpp"
+#include "duckdb/common/serializer/async_memory_governor.hpp"
 
 #include <functional>
 
@@ -20,7 +21,6 @@ namespace duckdb {
 
 class ClientContext;
 class TaskExecutor;
-class TemporaryMemoryState;
 
 //! Compile-time policy used by the async write layers.
 struct AsyncWriteConfig {
@@ -32,43 +32,8 @@ struct AsyncWriteConfig {
 	static constexpr idx_t LOCAL_COALESCE_THRESHOLD = 4096;
 	//! Remote file systems benefit from fewer round trips, so coalesce contiguous small buffers more aggressively.
 	static constexpr idx_t REMOTE_COALESCE_THRESHOLD = 8ULL * 1024ULL * 1024ULL;
-	//! Maximum queued async bytes retained per regular execution thread.
-	static constexpr idx_t MAX_PENDING_BYTES_PER_THREAD = 64ULL * 1024ULL * 1024ULL;
-	//! Minimum async write reservation requested per regular execution thread.
-	static constexpr idx_t MIN_PENDING_BYTES_PER_THREAD = 8ULL * 1024ULL * 1024ULL;
 	//! Maximum bytes a single managed stream request should submit before yielding scheduler capacity.
 	static constexpr idx_t DRAIN_TASK_BYTE_BUDGET = 16ULL * 1024ULL * 1024ULL;
-};
-
-//! Shared TemporaryMemoryManager reservation governor for the managed async queues.
-//! Encapsulates the coarse-growth reservation heuristic and backpressure budget so the positional write
-//! queue and the generic task queue bound their queued/in-flight backlog through identical memory logic.
-class ManagedAsyncMemoryGovernor {
-public:
-	explicit ManagedAsyncMemoryGovernor(ClientContext &client_context);
-
-	ManagedAsyncMemoryGovernor(const ManagedAsyncMemoryGovernor &) = delete;
-	ManagedAsyncMemoryGovernor &operator=(const ManagedAsyncMemoryGovernor &) = delete;
-
-	//! Whether a TemporaryMemoryState reservation is tracking this queue's backlog.
-	bool IsActive() const;
-	//! Grow the reservation coarsely until it covers current_pending_bytes; released only on Release().
-	void UpdateReservation(idx_t current_pending_bytes);
-	//! Current async backlog budget after applying the fixed cap, or 0 when memory is too tight to retain a backlog.
-	idx_t BackpressureBudget() const;
-	//! Release the reservation; further UpdateReservation calls may grow it again.
-	void Release();
-
-private:
-	ClientContext &client_context;
-	//! Temporary memory reservation state used to limit queued async data. Absent when draining synchronously.
-	unique_ptr<TemporaryMemoryState> memory_state;
-	//! Last remaining-size request sent to TemporaryMemoryManager. Grows monotonically until Release().
-	idx_t memory_request_bytes = 0;
-	//! Minimum TemporaryMemoryManager reservation while work is outstanding.
-	idx_t min_pending_bytes = 0;
-	//! Hard cap over the TemporaryMemoryState reservation.
-	idx_t max_pending_bytes = 0;
 };
 
 //! Owned payload that can be handed to an async write queue.
